python impl: remove old unused serdes routines
